backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
Faker==19.3.0
idna==3.4
iniconfig==2.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pluggy==1.0.0
psycopg==3.1.10
psycopg-binary==3.1.10
psycopg2-binary==2.9.9
Pygments==2.15.1
pytest==7.3.1
pytest-mock==3.10.0
pytest-print==0.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redgreenunittest==0.1.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
